dev-publications-timeline
* Fetch new publication data based on selected timeline
* Conditional render close and load more button on places page
