+++ server/controllers/ArtsController.ts
@@ -1,12 +1,13 @@
 import https from 'https';
-import { Request, Response, NextFunction } from 'express';
-
-import { addDataToRedis, getDataFromRedis } from '../redis';
-import { addDocument, elasticBulkCreate, elasticItemSearch } from '../elasticSearch';
 import { stackLogger } from 'info-logger';
+
+import { NextFunction, Request, Response } from 'express';
+
+import { addDocument, elasticBulkCreate, elasticItemSearch } from '../elasticSearch';
+import { addDataToRedis, getDataFromRedis } from '../redis';
 import {
-  placesData, healthCheckMessage, baseUrl, defaultEncoding,
-  HARVARD_API_KEY,
+  baseUrl, defaultEncoding, HARVARD_API_KEY, healthCheckMessage,
+  placesData,
 } from '../utils';
 
 class ArtsController {
