+++ src/Components/Collections/collections.scss
@@ -137,6 +137,7 @@
 
     &-header {
       padding-top: 100px;
+      padding-bottom: 30px;
       padding-left: 20px;
     }
