+++ src/index.scss
@@ -16,6 +16,10 @@ code {
     monospace;
 }
 
+h1, h2, h3, h4, h5, h6 {
+  color: unset;
+}
+
 .hide {
   display: none !important;
 }
