+++ tslint.json
@@ -16,7 +16,7 @@
     "no-conditional-assignment": true,
     "no-consecutive-blank-lines": false,
     "no-console": {
-      "severity": "warning",
+      "severity": "error",
       "options": [
         "debug",
         "info",
@@ -36,16 +36,9 @@
       2
     ],
     "interface-name": false,
-    "ordered-imports": false,
+    "ordered-imports": true,
     "object-literal-sort-keys": false,
     "jsx-wrap-multiline": true,
     "jsx-no-multiline-js": false
-  },
-  "jsRules": {
-    "max-line-length": {
-      "options": [
-        120
-      ]
-    }
   }
 }
